パッケージ名、ファイル構成の変更
