+++ src/main/java/com/springkadai/spring/Application.java
@@ -1,4 +1,4 @@
-package com.springkadai.spring.controller;
+package com.springkadai.spring;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
