+++ src/main/java/com/springkadai/spring/form/MovieCreateRequest.java
@@ -1,24 +1,25 @@
-package com.springkadai.spring.controller.response;
+package com.springkadai.spring.form;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDate;
 
-public class Movie {
-    private int movieId;
+public class MovieCreateRequest {
+    @NotBlank
     private String title;
+    @NotBlank
     private String filmDirector;
+    @NotBlank
     private String country;
+    @NotNull
     private LocalDate releaseDay;
 
-    public Movie(int movieId, String title, String filmDirector, String country, LocalDate releaseDay) {
-        this.movieId = movieId;
+    public MovieCreateRequest(String title, String filmDirector, String country, LocalDate releaseDay) {
         this.title = title;
         this.filmDirector = filmDirector;
         this.country = country;
         this.releaseDay = releaseDay;
-    }
-
-    public int getMovieId() {
-        return movieId;
     }
 
     public String getTitle() {
